+++ lib/src/revset.rs
@@ -1617,7 +1617,9 @@ pub fn walk_revs<'index>(
 
 fn resolve_git_ref(repo: &dyn Repo, symbol: &str) -> Option<Vec<CommitId>> {
     let view = repo.view();
-    for git_ref_prefix in &["", "refs/", "refs/heads/", "refs/tags/", "refs/remotes/"] {
+    // TODO: We should remove `refs/remotes` from this list once we have a better
+    // way to address local git repo's remote-tracking branches.
+    for git_ref_prefix in &["", "refs/", "refs/tags/", "refs/remotes/"] {
         if let Some(ref_target) = view.git_refs().get(&(git_ref_prefix.to_string() + symbol)) {
             return Some(ref_target.adds());
         }